Add main functionality of the reminder poll bot
This commit adds a SpringBoot application providing a telegram bot possibility to add, remove and list reminders. Each reminder is triggered on selected weeks, days, and at the selected time in the given chat. Each reminder is a poll proving chat users possibility to answer the poll question. The reminders are stored in memory